optimize code by splitting tests per different tabs
