+++ dcatestapp/build.gradle
@@ -22,7 +22,7 @@ android {
         minSdk 21
         targetSdk 34
         versionCode 1
-        versionName "3.5"
+        versionName "3.6"
 
         vectorDrawables {
             useSupportLibrary true
@@ -60,25 +60,13 @@ dependencies {
         changing = true
     }
 
+    implementation "androidx.viewpager2:viewpager2:1.1.0"
+    implementation "androidx.recyclerview:recyclerview:1.3.2"
+    implementation 'com.google.android.material:material:1.12.0'
     implementation "androidx.appcompat:appcompat:1.1.0"
 
     implementation 'com.google.code.gson:gson:2.8.2'
 
-    implementation 'io.reactivex.rxjava2:rxjava:2.2.19'
+    implementation 'io.reactivex.rxjava2:rxjava:2.2.21'
     implementation 'io.reactivex.rxjava2:rxandroid:2.1.1'
-
-    // JWT dependencies - if you want to parse JWTs
-    implementation 'net.jcip:jcip-annotations:1.0@jar'
-    implementation 'com.nimbusds:nimbus-jose-jwt:2.26@jar'
-    implementation 'net.minidev:json-smart:1.2@jar'
-
-//    implementation 'me.dm7.barcodescanner:zbar:1.8.4'
-//    implementation 'me.dm7.barcodescanner:zxing:1.8.4'
-
-    //compile 'com.google.zxing:zxing-parent:3.2.2-SNAPSHOT'
-    //compile(group: 'com.google.zxing', name: 'zxing-parent', version: '3.2.1', ext: 'pom')
-    implementation 'com.journeyapps:zxing-android-embedded:4.3.0@aar'
-    implementation 'com.google.zxing:core:3.4.1'
-
-//    implementation(platform("org.jetbrains.kotlin:kotlin-bom:1.9.23"))
 }
