+++ dcatestapp/src/main/AndroidManifest.xml
@@ -11,7 +11,7 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:theme="@style/AppTheme">
         <activity
-            android:name=".NonPaymentCardReaderActivity"
+            android:name=".ui.NonPaymentCardReaderActivity"
             android:exported="true"
             android:label="@string/app_name">
             <intent-filter>
