+++ dcatestapp/src/main/res/layout/fragment_mifare_tests.xml
@@ -1,0 +1,136 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:orientation="vertical">
+
+    <View
+        android:layout_width="match_parent"
+        android:layout_height="1dip"
+        android:background="#000000" />
+
+    <TextView
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="Mifare Classic Test" />
+
+    <HorizontalScrollView
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content">
+
+        <LinearLayout
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:orientation="horizontal">
+
+            <Button
+                android:id="@+id/testMifare"
+                style="@style/DCAButton"
+                android:text="Test Mifare" />
+
+            <Button
+                android:id="@+id/testMifareAfterPowerCycle"
+                style="@style/DCAButton"
+                android:text="Test Mifare (after power cycle)" />
+        </LinearLayout>
+
+    </HorizontalScrollView>
+
+    <View
+        android:layout_width="match_parent"
+        android:layout_height="1dip"
+        android:background="#000000" />
+
+    <TextView
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="Mifare Desfire Tests" />
+
+    <HorizontalScrollView
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content">
+
+        <LinearLayout
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:orientation="horizontal">
+
+            <Button
+                android:id="@+id/desfire601"
+                style="@style/DCAButton"
+                android:text="601"/>
+
+            <Button
+                android:id="@+id/desfire602"
+                style="@style/DCAButton"
+                android:text="602" />
+
+            <Button
+                android:id="@+id/desfire603"
+                style="@style/DCAButton"
+                android:text="603" />
+
+            <Button
+                android:id="@+id/desfire604"
+                style="@style/DCAButton"
+                android:text="604" />
+
+            <Button
+                android:id="@+id/desfire605"
+                style="@style/DCAButton"
+                android:text="605" />
+
+            <Button
+                android:id="@+id/desfire606"
+                style="@style/DCAButton"
+                android:text="606" />
+
+            <Button
+                android:id="@+id/desfire607"
+                style="@style/DCAButton"
+                android:text="607" />
+
+            <Button
+                android:id="@+id/desfire610"
+                style="@style/DCAButton"
+                android:text="610" />
+
+            <Button
+                android:id="@+id/desfire611"
+                style="@style/DCAButton"
+                android:text="611" />
+
+            <Button
+                android:id="@+id/desfire612"
+                style="@style/DCAButton"
+                android:text="612" />
+
+            <Button
+                android:id="@+id/desfire613"
+                style="@style/DCAButton"
+                android:text="613" />
+
+            <Button
+                android:id="@+id/desfire614"
+                style="@style/DCAButton"
+                android:text="614" />
+
+            <Button
+                android:id="@+id/desfire615"
+                style="@style/DCAButton"
+                android:text="615" />
+
+            <Button
+                android:id="@+id/desfire616"
+                style="@style/DCAButton"
+                android:text="616" />
+
+            <Button
+                android:id="@+id/desfire617"
+                style="@style/DCAButton"
+                android:text="617" />
+
+        </LinearLayout>
+    </HorizontalScrollView>
+
+</LinearLayout>
