+++ dcatestapp/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">DCA 3.5</string>
+    <string name="app_name">DCA 3.6</string>
 </resources>
