+++ dcatestapp/src/main/res/values/styles.xml
@@ -1,22 +1,12 @@
 <resources>
     <!-- Base application theme. -->
-    <style name="AppTheme" parent="@android:style/Theme.Holo.Light">
-        <!-- Customize your theme here.
-        <item name="android:windowActionBar">true</item>
-        <item name="android:windowNoTitle">true</item>
-        -->
-
-
+    <style name="AppTheme" parent="@style/Theme.AppCompat.Light">
     </style>
 
-    <style name="AppCompatCaptureTheme" parent="@android:style/Theme.Holo.Light">
-        <item name="android:windowFullscreen">true</item>
-        <item name="android:windowContentOverlay">@null</item>
-    </style>
     <style name="DCAButton">
         <item name="android:textSize">16sp</item>
-        <item name="android:minHeight">0dp</item>
-        <item name="android:maxWidth">150dp</item>
+        <item name="android:minHeight">50dp</item>
+        <item name="android:maxWidth">180dp</item>
         <item name="android:layout_width">wrap_content</item>
         <item name="android:layout_height">match_parent</item>
     </style>
